Generate secret-id if approle exists (#14)
Generate secret-id if approle exists

As a preparation step for giving various jobs possibility to access
vault base job verifies the project is authorized (requested role
exists) and it is running in post-review pipeline. When those conditions
are met allocate new secret-id and place it wrapped at the place where
job can take it.

Reviewed-by: Rodion Gyrbu <fpsoff@outlook.com>
Reviewed-by: Artem Goncharov <Artem.goncharov@gmail.com>